Chartのバージョン管理方法を改良
