+++ app/play/rhythmicalSlime.tsx
@@ -1,6 +1,7 @@
 "use client";
 import { useEffect, useRef, useState } from "react";
-import { BPMChange, Signature } from "@/../chartFormat/command.js";
+import { BPMChange } from "@/../chartFormat/bpm.js";
+import { Signature } from "@/../chartFormat/signature.js";
 import { getSignatureState, getTimeSec } from "@/../chartFormat/seq.js";
 import { Step, stepAdd, stepSub, stepZero } from "@/../chartFormat/step.js";
 import { useDisplayMode } from "@/scale.js";
