+++ chartFormat/legacy/chart2.ts
@@ -1,18 +1,12 @@
-import { emptyLevel } from "../chart.js";
-import { BPMChange } from "../command.js";
-import { luaAddBpmChange } from "../lua/bpm.js";
-import { luaAddNote } from "../lua/note.js";
-import { luaAddSpeedChange } from "../lua/speed.js";
-import { findBpmIndexFromStep, getTimeSec } from "../seq.js";
-import { Step, stepZero } from "../step.js";
-import { Chart3 } from "./chart3.js";
+import { stepZero } from "../step.js";
+import { BPMChange1, Chart1, NoteCommand1 } from "./chart1.js";
 
 export interface Chart2 {
   falling: "nikochan"; // magic
   ver: 2;
-  notes: NoteCommand2[];
-  bpmChanges: BPMChange[];
-  scaleChanges: BPMChange[];
+  notes: NoteCommand1[];
+  bpmChanges: BPMChange1[];
+  scaleChanges: BPMChange1[];
   offset: number;
   ytId: string;
   title: string;
@@ -20,99 +14,26 @@ export interface Chart2 {
   chartCreator: string;
   editPasswd: string;
 }
-export interface NoteCommand2 {
-  step: Step;
-  big: boolean;
-  hitX: number;
-  hitVX: number;
-  hitVY: number;
-  accelY: number;
-}
 
-export function convert2To3(chart: Chart2): Chart3 {
-  let newChart: Chart3 = {
+export function convert1To2(chart: Chart1): Chart2 {
+  return {
     falling: "nikochan",
-    ver: 3,
-    notes: [],
-    rest: [],
-    bpmChanges: [],
-    speedChanges: [],
+    ver: 2,
+    notes: chart.notes.map((n) => ({
+      step: n.step,
+      big: n.big,
+      hitX: n.hitX * 10 - 5,
+      hitVX: n.hitVX * 4,
+      hitVY: n.hitVY * 4,
+      accelY: n.accelY * 4,
+    })),
+    bpmChanges: chart.bpmChanges,
+    scaleChanges: [{ step: stepZero(), timeSec: 0, bpm: 120 }],
     offset: chart.offset,
-    lua: [],
     ytId: chart.ytId,
     title: chart.title,
     composer: chart.composer,
     chartCreator: chart.chartCreator,
     editPasswd: chart.editPasswd,
   };
-  for (const n of chart.bpmChanges) {
-    newChart = {
-      ...newChart,
-      ...luaAddBpmChange({ ...emptyLevel(), ...newChart }, n)!,
-    };
-  }
-  if (chart.bpmChanges.length === 0) {
-    newChart = {
-      ...newChart,
-      ...luaAddBpmChange(
-        { ...emptyLevel(), ...newChart },
-        {
-          bpm: 120,
-          step: stepZero(),
-          timeSec: 0,
-        }
-      )!,
-    };
-  }
-  for (const n of chart.scaleChanges) {
-    newChart = {
-      ...newChart,
-      ...luaAddSpeedChange({ ...emptyLevel(), ...newChart }, n)!,
-    };
-  }
-  if (chart.scaleChanges.length === 0) {
-    newChart = {
-      ...newChart,
-      ...luaAddSpeedChange(
-        { ...emptyLevel(), ...newChart },
-        {
-          bpm: 120,
-          step: stepZero(),
-          timeSec: 0,
-        }
-      )!,
-    };
-  }
-  for (const n of chart.notes) {
-    const nScale2 =
-      chart.scaleChanges[findBpmIndexFromStep(chart.scaleChanges, n.step)].bpm;
-    const currentSpeed3 =
-      newChart.speedChanges[findBpmIndexFromStep(newChart.speedChanges, n.step)]
-        .bpm;
-    const newSpeed3 = Math.sqrt(nScale2 * nScale2 * n.accelY);
-    const newVX = (nScale2 * n.hitVX) / newSpeed3;
-    const newVY = (nScale2 * n.hitVY) / newSpeed3;
-    if (currentSpeed3 !== newSpeed3) {
-      newChart = {
-        ...newChart,
-        ...luaAddSpeedChange(
-          { ...emptyLevel(), ...newChart },
-          {
-            bpm: newSpeed3,
-            step: n.step,
-            timeSec: getTimeSec(newChart.bpmChanges, n.step),
-          }
-        )!,
-      };
-    }
-    newChart = {
-      ...newChart,
-      ...luaAddNote(
-        { ...emptyLevel(), ...newChart },
-        { hitX: n.hitX, hitVX: newVX, hitVY: newVY, step: n.step, big: n.big },
-        n.step
-      )!,
-    };
-  }
-  return newChart;
 }
