+++ chartFormat/legacy/chart7.ts
@@ -1,10 +1,10 @@
 import { BPMChangeWithLua3, NoteCommandWithLua3, RestStep3 } from "./chart3.js";
-import { Chart5, SignatureWithLua5 } from "./chart5.js";
+import { SignatureWithLua5 } from "./chart5.js";
 
-export interface Chart6 {
+export interface Chart7 {
   falling: "nikochan"; // magic
-  ver: 6;
-  levels: Level6[];
+  ver: 7;
+  levels: Level7[];
   offset: number;
   ytId: string;
   title: string;
@@ -13,7 +13,7 @@ export interface Chart6 {
   editPasswd: string;
   published: boolean;
 }
-export interface Level6 {
+export interface Level7 {
   name: string;
   type: string;
   notes: NoteCommandWithLua3[];
@@ -23,16 +23,4 @@ export interface Level6 {
   signature: SignatureWithLua5[];
   lua: string[];
   unlisted: boolean;
-}
-
-export function convert5To6(chart: Chart5): Chart6 {
-  return {
-    ...chart,
-    levels: chart.levels.map((level) => ({
-      ...level,
-      unlisted: !!level.unlisted,
-    })),
-    ver: 6,
-    published: false,
-  };
 }
