+++ chartFormat/seq.ts
@@ -1,6 +1,14 @@
 import { Chart } from "./chart.js";
-import { BPMChange, getBarLength, Signature, toStepArray } from "./command.js";
-import { Step, stepAdd, stepCmp, stepSub, stepToFloat, stepZero } from "./step.js";
+import { getBarLength, Signature, toStepArray } from "./signature.js";
+import { BPMChange } from "./bpm.js";
+import {
+  Step,
+  stepAdd,
+  stepCmp,
+  stepSub,
+  stepToFloat,
+  stepZero,
+} from "./step.js";
 
 export interface ChartSeqData {
   notes: Note[];
